サウンド修正

+++ source/music.h
@@ -88,7 +88,7 @@ struct ds_struc
 GLOBAL struct ds_struc DSmusic;	// 楽曲用
 GLOBAL struct ds_struc DSeffect;	// 効果音用
 
-GLOBAL u32 save_REG_SGCNT0_H;
+GLOBAL u16 save_REG_SGCNT0_H;
 
 #define PLAY_LOOP_ON  1 // ループ演奏 1:ループ
 #define PLAY_LOOP_OFF 0 // ループ演奏 0:ループ無し
